feat(WebAuthn): support and default to 'none' attestation (#19)
* [Example] Add a development workflow

Add a script to the example app that runs the server in port :3001, and
proxy requests from `react-scripts start` to it. This allows running the
example in an unoptimized mode suitable for development.

Fixes #16

* Support and default to none attestation

Add support for 'none' attestation. Consider it valid if the webauthn
config sets the conveyance preference to none.

This change also shuffles some code around to avoid repetition when
verifying the MakeCredential attestation.

Fixes #18

* Add docs.